Module core_foundation_sys::base

Structs

CFAllocatorContext

CFRange

Enums

CFComparisonResult

Statics

kCFAllocatorDefault

kCFAllocatorMalloc

kCFAllocatorMallocZone

kCFAllocatorNull

kCFAllocatorSystemDefault

kCFAllocatorUseContext

kCFNotFound

Constant used by some functions to indicate failed searches.

kCFNull

Traits

TCFTypeRef

Trait for all types which are Core Foundation reference types.

Functions

CFAllocatorAllocate⚠

CFAllocatorCreate⚠

CFAllocatorDeallocate⚠

CFAllocatorGetContext⚠

CFAllocatorGetDefault⚠

CFAllocatorGetPreferredSizeForSize⚠

CFAllocatorGetTypeID⚠

CFAllocatorReallocate⚠

CFAllocatorSetDefault⚠

CFCopyDescription⚠

CFEqual⚠

CFGetRetainCount⚠

CFGetTypeID⚠

CFHash⚠

CFRelease⚠

CFRetain⚠

CFShow⚠

Type Definitions

Boolean

CFAllocatorAllocateCallBack

CFAllocatorCopyDescriptionCallBack

CFAllocatorDeallocateCallBack

CFAllocatorPreferredSizeCallBack

CFAllocatorReallocateCallBack

CFAllocatorRef

CFAllocatorReleaseCallBack

CFAllocatorRetainCallBack

CFHashCode

CFIndex

CFNullRef

CFOptionFlags

CFTypeID

CFTypeRef

OSStatus

SInt32

mach_port_t