core_foundation_sys

Module base

Structs

CFAllocatorContext

CFRange

Enums

CFComparisonResult

Statics

kCFAllocatorDefault⚠

kCFAllocatorMalloc⚠

kCFAllocatorMallocZone⚠

kCFAllocatorNull⚠

kCFAllocatorSystemDefault⚠

kCFAllocatorUseContext⚠

kCFNotFound

Constant used by some functions to indicate failed searches.

kCFNull⚠

Traits

TCFTypeRef

Trait for all types which are Core Foundation reference types.

Functions

CFAllocatorAllocate⚠

CFAllocatorCreate⚠

CFAllocatorDeallocate⚠

CFAllocatorGetContext⚠

CFAllocatorGetDefault⚠

CFAllocatorGetPreferredSizeForSize⚠

CFAllocatorGetTypeID⚠

CFAllocatorReallocate⚠

CFAllocatorSetDefault⚠

CFCopyDescription⚠

CFCopyTypeIDDescription⚠

CFEqual⚠

CFGetAllocator⚠

CFGetRetainCount⚠

CFGetTypeID⚠

CFHash⚠

CFNullGetTypeID⚠

CFRelease⚠

CFRetain⚠

CFShow⚠

Type Aliases

Boolean

CFAllocatorAllocateCallBack

CFAllocatorCopyDescriptionCallBack

CFAllocatorDeallocateCallBack

CFAllocatorPreferredSizeCallBack

CFAllocatorReallocateCallBack

CFAllocatorRef

CFAllocatorReleaseCallBack

CFAllocatorRetainCallBack

CFComparatorFunction

CFHashCode

CFIndex

CFNullRef

CFOptionFlags

CFTypeID

CFTypeRef

ConstStr255Param

ConstStringPtr

LangCode

OSStatus

RegionCode

SInt16

SInt32

StringPtr

UInt8

UInt16

UInt32

UTF8Char

UTF16Char

UTF32Char

mach_port_t